[dynamo] Possibly handle UserFunctionVariable as a TorchVariable
In some cases, a UserFunctionVariable can be constructed when the underlying function is actually a TorchVariable. One example is when an attribute on a UnspecializedNNModuleVariable is a torch function. In those cases, we should treat the UserFunctionVariable as a TorchVariable.

This adds a check in UserFunctionVariable.call_function: if self.fn is_allowed, then wrap it in a TorchVariable and call_function on that.

Fixes #104172

ghstack-source-id: 87c7c2afaddc2c4dcf6543880c8a391950a7ad3a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/104231